server/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
dnspython==2.7.0
email_validator==2.2.0
environs==14.2.0
fastapi==0.116.1
fastapi_cors==0.0.6
h11==0.16.0
idna==3.10
marshmallow==4.0.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
azure-ai-documentintelligence
pandas
dotenv